split off schemaorg_author and bibtex_author into their own files
